+++ src/Faker.Tests/Extensions/StringExtensions.cs
@@ -2,12 +2,16 @@
 
 namespace RimuTec.Faker.Tests.Extensions {
    static class StringExtensions {
-      public static string RemoveFullStops(this string source) {
+      public static string RemovePeriods(this string source) {
          return source.Replace(".", string.Empty);
       }
 
+      public static string RemoveQuestionMarks(this string source) {
+         return source.Replace("?", string.Empty);
+      }
+
       public static IEnumerable<string> ToWordList(this string source) {
-         return source.RemoveFullStops().ToLower().Split(' ');
+         return source.RemovePeriods().RemoveQuestionMarks().ToLower().Split(' ');
       }
    }
 }