+++ src/Faker.Tests/Extensions/StringExtensionsTests.cs
@@ -14,5 +14,17 @@ namespace RimuTec.Faker.Tests.Extensions {
          // assert
          Assert.AreEqual("must||be||equal||to||or||greater||than||zero", wordList);
       }
+
+      [Test]
+      public void ToWordList_For_Question() {
+         // arrange
+         var sentence = "What is the answer to all questions?";
+
+         // act
+         var wordList = string.Join("||", sentence.ToWordList());
+
+         // assert
+         Assert.AreEqual("what||is||the||answer||to||all||questions", wordList);
+      }
    }
 }